==> LASI/1_Sociodemo_econom/1_2_mbirth.Rmd <==
---
title: "Month of Birth"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 
<!-- # Installation, loading packages -->
```{r setup, include=FALSE}
source("../../setup.r")
```
<!-- ########################################################## --> 
<!-- # Loading DS description -->
```{r main, child='../../_DS_Rmd/mbirthDS.Rmd'}
```
```{r global, echo=F}
source('../../_DS_Rmd/mbirthDS.R')
```


# Data process

## India

### Wave 1 

#### Study-specific variable description

| **Name** | `dm007_month` |
|----------|------------|
| **Label** | `date of birth month` |
| **Table name** | `LASI-Pilot_all` |
| **Categories** |`1 = january; 2 = february; 3 = march; 4 = april; 5 = may; 6 = june; 7 = july; 8 = august; 9 = september; 10 = october; 11 = november; 12 = december; 13 = month not known`|
| **Description** |  |


```{r assign1, echo=F}
opal.assign.table.tibble(o, 'LASI_w1_mbirth','LASI.LASI-Pilot_all',
                         variables=list('dm007_month'), missings = TRUE)
LASI_w1_mbirth <- opal.execute(o,'LASI_w1_mbirth')
```

```{r local1, echo=F}
vari <- LASI_w1_mbirth$dm007_month
kable(Categorical_summary(vari, missing_values = NA)[2], caption = "dm007_month") 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(LASI_w1_mbirth, aes(x=factor(dm007_month))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vari)) + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `recode 13 "month not known" into 998`
* `NA into 999`

**R script:**

```{r harmo1}
LASI_ds_w1 <- tibble(id=LASI_w1_mbirth$id)
LASI_ds_w1$mbirth <- car::recode(LASI_w1_mbirth$dm007_month, "NA=999; 13=998")
```

#### Statistics of the new harmonised variable
```{r descript1, echo=F}
vari <- LASI_ds_w1$mbirth
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(LASI_ds_w1, aes(x=factor(mbirth))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Month of birth") + ylab("Frequency")
```

#### Validation
```{r crosstabulation1, echo=F}
BeforeH <- c(table(LASI_w1_mbirth$dm007_month), sum(is.na(LASI_w1_mbirth$dm007_month)))
AfterH <- table(LASI_ds_w1$mbirth)
C <- rbind(BeforeH,AfterH)
colnames(C) <- c(names(AfterH)[1:12], "13->998", "NA->999")
kable(C)
```


```{r importOPAL1, echo=F}

#opal.assign.data(o, 'LASI_ds_w1', LASI_ds_w1)
#opal.symbol_import(o,'LASI_ds_w1', project='_Athlos_Harm_Dataset')
mbirth <- LASI_ds_w1
save(mbirth, file = "../RData/w1/mbirth.RData")
rm(mbirth)


LASI_m_ds_w1 <- LASI_ds_w1
LASI_m_ds_w1$mbirth <- car::recode(LASI_m_ds_w1$mbirth, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'LASI_m_ds_w1', LASI_m_ds_w1)
#opal.symbol_import(o,'LASI_m_ds_w1', project='_Athlos_Harm_Dataset')
```



# Quality estimation
No comments.


```{r closeRsession, echo=FALSE}
opal.logout(o)
```
