==> LASI/1_Sociodemo_econom/1_3_ybirth.Rmd <==
---
title: "Year of Birth"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 
<!-- # Installation, loading packages -->
```{r setup, include=FALSE}
source("../../setup.r")
```
<!-- ########################################################## --> 
<!-- # Loading DS description -->
```{r main, child='../../_DS_Rmd/ybirthDS.Rmd'}
```
```{r global, echo=F}
source('../../_DS_Rmd/ybirthDS.R')
```

# Data process

## India

### Wave 1 

#### Study-specific variable description

| **Name** | `dm007_year` |`dm008`|`yintw`|
|--|--------|-------|-----|
| **Label** | `date of birth year` | `age in completed years`|`Year of interview`|
| **Table name** | `LASI-Pilot_all` | `LASI-Pilot_all` |`lasi_pilot_hd`|
| **Description** |  |  |  |



```{r assign1, echo=F}
opal.assign.table.tibble(o, 'LASI_w1_ybirth','LASI.LASI-Pilot_all',
                         variables=list('dm007_year','dm008'), missings = TRUE)
LASI_w1_ybirth <- opal.execute(o,'LASI_w1_ybirth')

opal.assign.table.tibble(o, 'LASI_w1','LASI_HD.lasi_pilot_hd',
                         variables=list('yintw'), missings = TRUE)
LASI_w1 <- opal.execute(o,'LASI_w1')

LASI_w1_ybirth <- left_join(LASI_w1_ybirth,LASI_w1, by="id")

```

```{r local1, echo=F}
vari <- LASI_w1_ybirth$dm007_year
kable(Continuous_summary(vari, missing_values = NA)[3], caption = "dm007_year") 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(LASI_w1_ybirth, aes(dm007_year)) + geom_histogram(stat="count", fill="steelblue") + xlab("Year of birth") + ylab("Frequency")

vari <- LASI_w1_ybirth$dm008
kable(Continuous_summary(vari, missing_values = NA)[3], caption = "dm008")
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(LASI_w1_ybirth, aes(dm008)) + geom_histogram(stat="count", fill="steelblue") + xlab(label(vari)) + ylab("Frequency")

vari <- LASI_w1_ybirth$yintw
kable(Categorical_summary(vari, missing_values = NA)[2], caption = "yintw")
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Values")
ggplot(LASI_w1_ybirth, aes(x=factor(yintw))) + geom_histogram(stat="count", fill="steelblue") + xlab(label(vari)) + ylab("Frequency")

```


#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable `dm007_year`, but the missing values have to be recoded as follows:
There is a value 2010, It is going to be consider as missing, recoded as 999. 
Other missing values will be approximated by age and year of interview.

* `NA into 999`

**R script:**

```{r harmo1}
LASI_ds_w1 <- tibble(id=LASI_w1_ybirth$id)
LASI_ds_w1$ybirth <- car::recode(LASI_w1_ybirth$dm007_year, "NA=999; 2010=999")

LASI_w1_ybirth$cage <- LASI_w1_ybirth$yintw - LASI_w1_ybirth$dm008

LASI_ds_w1$ybirth[which(LASI_ds_w1$ybirth %in% miss_values_vector)] <- car::recode(LASI_w1_ybirth$cage[which(LASI_ds_w1$ybirth %in% miss_values_vector)], "NA=999")

```

#### Statistics of the new harmonised variable
```{r descript1, echo=F}
vari <- LASI_ds_w1$ybirth
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = "Year of birth")
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(LASI_ds_w1, aes(ybirth)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vari %in% miss_values_vector)) + xlab("Year of birth") + ylab("Frequency")
```

#### Validation
```{r crosstabulation1, echo=F}
```



```{r importOPAL1, echo=F}
ybirth <- LASI_ds_w1
save(ybirth, file = "../RData/w1/ybirth.RData")
rm(ybirth)

```


# Quality estimation
No comments.


```{r closeRsession, echo=FALSE}
opal.logout(o)
```
