==> LASI/1_Sociodemo_econom/2_age.Rmd <==
---
title: "Age"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 
<!-- # Installation, loading packages -->
```{r setup, include=FALSE}
source("../../setup.r")
```
<!-- ########################################################## --> 
<!-- # Loading DS description -->
```{r main, child='../../_DS_Rmd/ageDS.Rmd'}
```
```{r global, echo=F}
source('../../_DS_Rmd/ageDS.R')
```


# Data process

## India


### Wave 1 

#### Study-specific variable description

| **Name** | `dm008` |`mbirth`|`ybirth`|
|-|-|-|-|
| **Label** | `age in completed years` |`Month of birth`|`Year of birth`|
| **Table name** | `LASI-Pilot_all` |`mbirth`|`ybirth`|
| **Missings** | `NA` |`998 = Do not know`<br/>`999 = Missing`|`999 = Missing`|
| **Description** |  |  |  |


```{r assign1, echo=F}

opal.assign.table.tibble(o, 'LASI_w1','LASI.LASI-Pilot_all', variables=list('dm008'), missings = TRUE)
LASI_w1 <- opal.execute(o,'LASI_w1')
load(paste0(datafolder,"w1/mbirth.RData"))
LASI_w1 <- left_join(LASI_w1,mbirth,by="id")
rm(mbirth)
load(paste0(datafolder,"w1/ybirth.RData"))
LASI_w1 <- left_join(LASI_w1,ybirth,by="id")
rm(ybirth)

```

```{r local1, echo=F}

vari <- LASI_w1$dm008
kable(Continuous_summary(vari, missing_values = NA)[3], caption = "dm008")
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(LASI_w1, aes(dm008)) + geom_histogram(stat="count", fill="steelblue") + xlab(label(vari)) + ylab("Frequency")

vari <- LASI_w1$mbirth
kable(Continuous_summary(vari, missing_values = miss_values_vector)[3], caption = "mbirth")
pander(Continuous_summary(vari, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(LASI_w1, aes(mbirth)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vari %in% miss_values_vector)) + xlab(label(vari)) + ylab("Frequency")

vari <- LASI_w1$ybirth
kable(Continuous_summary(vari, missing_values = miss_values_vector)[3], caption = "ybirth")
pander(Continuous_summary(vari, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(LASI_w1, aes(ybirth)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vari %in% miss_values_vector)) + xlab(label(vari)) + ylab("Frequency")

```

#### Harmonisation algorithm
The harmonised variable is the same than the study-specific variable `dm008`, but the missing values have to be recoded as follows:

* `for non-missing ybirth, age = 2010 (yintw) - ybirth + (6-mbirth)/12`
* `otherwise, NA into 999`

**R script:**

```{r harmo1}

LASI_ds_w1 <- tibble(id=LASI_w1$id)
LASI_ds_w1$age <- car::recode(LASI_w1$dm008, "NA=999")
LASI_w1$mbirth <- car::recode(LASI_w1$mbirth,"miss_values_vector=7")
LASI_ds_w1$age[which(LASI_ds_w1$age %in% miss_values_vector & !LASI_w1$ybirth %in% miss_values_vector)] <- floor((2010-LASI_w1$ybirth[which(LASI_ds_w1$age %in% miss_values_vector & !LASI_w1$ybirth %in% miss_values_vector)])+(6-LASI_w1$mbirth[which(LASI_ds_w1$age %in% miss_values_vector & !LASI_w1$ybirth %in% miss_values_vector)])/12)

```

#### Statistics of the new harmonised variable
```{r descript1, echo=F}
vari <- LASI_ds_w1$age
kable(Continuous_summary(var = vari, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = miss_values_vector)[2], caption = "values")
pander(Continuous_summary(vari, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(LASI_ds_w1, aes(age)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vari %in% miss_values_vector)) + xlab(ds_label) + ylab("Frequency")
```

#### Validation
```{r crosstabulation1, echo=F}

load(paste0(datafolder,"w1/yintw",".RData"))
LASI_w1 <- left_join(LASI_w1,yintw, by="id")
LASI_w1$ybirth <- car::recode(LASI_w1$ybirth, "miss_values_vector=NA")
LASI_w1$yintw <- car::recode(LASI_w1$yintw, "miss_values_vector=NA")
LASI_w1$cage <- LASI_w1$yintw-LASI_w1$ybirth
LASI_w1$diff <- abs(LASI_w1$dm008 - LASI_w1$cage)
LASI_w1 <- remove_val_labels(LASI_w1)
kable(frq(LASI_w1$diff), caption = 'Frequencies of differences between harmonised age and age computed as year of interview - year of birth')
kable(LASI_w1[which(LASI_w1$diff>2),], align = 'c', caption = 'There is an individuals with difference greater than 2')

```



```{r importOPAL1, echo=F}


LASI_ds_w1$age <- labelled(LASI_ds_w1$age, labels = cont_label)
label(LASI_ds_w1$age) <- ds_label

#opal.assign.data(o, 'LASI_ds_w1', LASI_ds_w1)
#opal.symbol_import(o,'LASI_ds_w1', project='_Athlos_Harm_Dataset')
age <- LASI_ds_w1
save(age, file = "../RData/w1/age.RData")
rm(age)

LASI_m_ds_w1 <- LASI_ds_w1
LASI_m_ds_w1$age <- car::recode(LASI_m_ds_w1$age, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'LASI_m_ds_w1', LASI_m_ds_w1)
#opal.symbol_import(o,'LASI_m_ds_w1', project='_Athlos_Harm_Dataset')
```


# Quality estimation
No comments.


```{r closeRsession, echo=FALSE}
opal.logout(o)
```
