==> LASI/1_Sociodemo_econom/3_sex.Rmd <==
---
title: "Sex"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 
<!-- # Installation, loading packages -->
```{r setup, include=FALSE}
source("../../setup.r")
```
<!-- ########################################################## --> 
<!-- # Loading DS description -->
```{r main, child='../../_DS_Rmd/sexDS.Rmd'}
```
```{r global, echo=F}
source('../../_DS_Rmd/sexDS.R')
```




<!-- ########################################################## --> 

# Data process

## India

### Wave 1 

#### Study-specific variable description


| **Name** | `dm002`|
|-|-|
| **Label** | `gender of respondent`|
| **Table name** |`LASI-Pilot_all`|
| **Categories** |`1 = Male`<br/>`2 = Female`|
| **Description** |  |


```{r assign1, echo=FALSE}
opal.assign.table.tibble(o, 'LASI_w1_sex','LASI.LASI-Pilot_all',
                         variables=list('dm002'), missings = TRUE)
LASI_w1_sex <- opal.execute(o,'LASI_w1_sex')
```

```{r local1, echo=FALSE}
vari <- LASI_w1_sex$dm002
kable(Categorical_summary(vari, missing_values = NA)[2], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(LASI_w1_sex, aes(x=factor(dm002))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vari)) + ylab("Frequency")
```

#### Harmonisation algorithm
To compute the harmonized variable from the study-specific variable it has to be recoded as follows:

* `2 "Female" into 0`
* `keep "Male" into 1`

**R script:**

```{r harmo1, echo=TRUE}
LASI_ds_w1 <- tibble(id=LASI_w1_sex$id)
LASI_ds_w1$sex <- car::recode(LASI_w1_sex$dm002, "2=0; NA=999")
LASI_ds_w1$sex <- labelled(LASI_ds_w1$sex, labels = c("Female"=0, "Male"=1, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript1, echo=FALSE}
vari <- LASI_ds_w1$sex
kable(Categorical_summary(vari, missing_values = miss_values_vector)[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(LASI_ds_w1, aes(x=factor(sex))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Validation
```{r crosstabulation1, echo=FALSE}
AfterH <- table(LASI_ds_w1$sex)
BeforeH <- table(LASI_w1_sex$dm002)[c(2,1)]
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("2->0","1")
kable(C)
```





```{r helpfunctions, echo=F}
source("../../helpfunctions.r")
```



```{r summ, echo=F}

# All study waves and populations with abbreviated and descriptive names


lasi.cw <- list(w1 = c("w1","W1"))

# Consider only harmonised waves
l.hds <- lapply(lasi.cw, function(wname) if(exists(paste0("LASI_ds_",wname[1]))){wname = list(get(paste0("LASI_ds_",wname[1])),wname[2])})
nullw <- which(sapply(l.hds, is.null))

if(length(nullw)!=0){
  l.hds <- l.hds[-nullw]
}

m.hds <- list()

lmlist <- list(lapply(l.hds,function(x)x[[1]]),m.hds)

if(length(l.hds)>0){
  # Labelling tibbles
  lmlist <- labelling(lmlist[[1]],lmlist[[2]])
  # Printing table of categories
  # print(knitr::kable(summaries(lmlist[[1]],lmlist[[2]],lapply(l.hds,function(x)x[[2]])),caption=ds_label))
  # Printing table of trajectories
  # f.dbb <- trajectories(lmlist[[2]],vbl_name=hd_vbl)
  # print(knitr::kable(f.dbb))
  # Saving tibbles
  savingRD(lmlist[[1]],vbl_name=hd_vbl)

}
```


# Quality estimation

No comments.


```{r closeRsession, echo=FALSE} 
opal.logout(o)
```
