==> LASI/1_Sociodemo_econom/4_marital_status.Rmd <==
---
title: "Marital Status"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 
<!-- # Installation, loading packages -->
```{r setup, include=FALSE}
source("../../setup.r")
```
<!-- ########################################################## --> 
<!-- # Loading DS description -->
```{r main, child='../../_DS_Rmd/marital_statusDS.Rmd'}
```
```{r global, echo=F}
source('../../_DS_Rmd/marital_statusDS.R')
```



<!-- ########################################################## --> 

# Data process
## India
### Wave 1 

#### Study-specific variable description

 
| **Name** | `rmaritalstatus_r` |
|-|-|
| **Label** | `r '\x60r marital status \x60'` |
| **Table name** | `LASI-Pilot_all` |
| **Categories:**| `1 = never married`<br/>`2 = currently married`<br/>`3 = separated/deserted`<br/>`4 = divorced`<br/>`5 = widowed` |
| **Description** |  | 


```{r assign1,  echo=FALSE}

opal.assign.table.tibble(o, 'LASI_w1_ms','LASI.LASI-Pilot_all',variables=list('rmaritalstatus_r'), missings = TRUE)
LASI_w1_ms <- opal.execute(o,'LASI_w1_ms')

```


```{r local1,  echo=FALSE}
vari <- LASI_w1_ms$rmaritalstatus_r
kable(Categorical_summary(vari, missing_values = NA)[2], caption = "rmaritalstatus_r") 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
ggplot(LASI_w1_ms, aes(x=factor(rmaritalstatus_r))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vari)) + ylab("Frequency")
```

#### Harmonisation algorithm

To compute the harmonized variable from the study-specific variable it has to be recoded as follows:

* `keep "never married" into 1`
* `keep "currently married" into 2`
* `keep "separated/deserted" into 3`
* `recode 4 "divorced" into 3`
* `recode 5 "widowed" into 4`

**R script:**

```{r harmo1,   echo=TRUE}
LASI_ds_w1 <- tibble(id=LASI_w1_ms$id)
LASI_ds_w1$marital_status <- car::recode(LASI_w1_ms$rmaritalstatus_r, "4=3; 5=4; NA=999")
LASI_ds_w1$marital_status <- labelled(LASI_ds_w1$marital_status, labels = c("single"=1, "married cohabitating"=2, "divorced separated"=3, "widow"=4, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript1,  echo=FALSE}
vari <- LASI_ds_w1$marital_status
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3]) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(LASI_ds_w1, aes(x=factor(marital_status))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("marital status") + ylab("Frequency")
```

#### Validation
```{r crosstabulation1,  echo=FALSE}
BeforeH <- c(table(LASI_w1_ms$rmaritalstatus_r))
BeforeH <- c(BeforeH[c(1,2)], sum(BeforeH[c(3, 4)]), BeforeH[5] )
AfterH <- table(LASI_ds_w1$marital_status)
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("1","2", "(3,4)->3", "5->4")
kable(C)
```


```{r importOPAL1,  echo=FALSE}

marital_status <- LASI_ds_w1
save(marital_status, file = "../RData/w1/marital_status.RData")
rm(marital_status)

```


# Quality estimation

No comments


```{r closeRsession, echo=FALSE,  echo=FALSE} 
opal.logout(o)
```
