==> LASI/1_Sociodemo_econom/5_education.Rmd <==
---
title: "Education"
output: 
  html_document:
    theme: united
    highlight: tango
    toc: true
    toc_depth: 6
    toc_float:
      collapsed: true
      smooth_scroll: true
    number_sections: true
---

<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->



```{r setup, include=FALSE,  echo=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_ania.r")
```

<!-- ########################################################## --> 

# Description of DataSchema variable 

* Short name: `education`
* Variable label: `Highest level of formal education achieved`
* Domain: `Socio-demographic and economic characteristics`
* Value type: `categorical`
* Category coding:

**Code** | **Category Label**
-------- | ------------------
0        | less than primary education
1        | primary education
2        | secondary education
3        | tertiary education

* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing




<!-- ########################################################## --> 

# Data process
   
## India
### Wave 1 

#### Study-specific variable description

 


* Name: `dm031_n_edu`
* Label: `level of education`
* Categories: 
    + `0 = no schooling`
    + `1 = less than primary`
    + `2 = primary completed`
    + `3 = middle school completed`
    + `4 = high school completed`
    + `5 = more than high school`

* Description:
```{r assign1,  echo=FALSE}
opal.assign.table.tibble(o, 'LASI_w1_educ','LASI.LASI-Pilot_all',
                         variables=list('dm031_n_edu'), missings = TRUE)
```

```{r local1,  echo=FALSE}
LASI_w1_educ <- opal.execute(o,'LASI_w1_educ')
vari <- LASI_w1_educ$dm031_n_edu
kable(Categorical_summary(vari, missing_values = NA)[2], caption = attributes(vari)$`spss::shortName`) 
kable(Categorical_summary(vari, missing_values = NA)[1], caption = "Category")
#Categorical_summary(var = LASI_w1_educ$dm031_n_edu, missing_values = NA)
ggplot(LASI_w1_educ, aes(x=factor(dm031_n_edu))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Level of education") + ylab("Frequency")
```

#### Harmonisation algorithm

To compute the harmonized variable from the study-specific variable it has to be recoded as follows:

* `keep "no schooling" into 0`
* `recode 1 "less than primary" into 0`
* `recode 2 "primary completed" into 1`
* `recode 3 "middle school completed" into 2`
* `recode 4 "high school completed" into 2`
* `recode 5 "more than high school" into 3`

**R script:**

```{r harmo1,  echo=TRUE}
LASI_ds_w1 <- tibble(id=LASI_w1_educ$id)
LASI_ds_w1$education <- car::recode(LASI_w1_educ$dm031_n_edu, "1=0; 2=1; 3=2; 4=2; 5=3;  NA=999")
LASI_ds_w1$education <- labelled(LASI_ds_w1$education, labels = c("less than primary education"=0, "primary education"=1, "secondary education"=2, "tertiary education"=3, "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript1,  echo=FALSE}
vari <- LASI_ds_w1$education
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3]) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
#Categorical_summary(var = LASI_ds_w1$education, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(LASI_ds_w1, aes(x=factor(education))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Level of education") + ylab("Frequency")
```

#### Validation
```{r crosstabulation1,  echo=FALSE}
BeforeH <- c(table(LASI_w1_educ$dm031_n_edu), sum(is.na(LASI_w1_educ$dm031_n_edu)))
BeforeH <- c(sum(BeforeH[c(1,2)]),BeforeH[3], sum(BeforeH[c(4,5)]), BeforeH[6], BeforeH[7] )
AfterH <- table(LASI_ds_w1$education)
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("(0,1)->0",  "1", "(3,4)->2", "5->3", "NA->999")
kable(C)
```


```{r importOPAL1,  echo=FALSE}
val_labels(LASI_ds_w1) <- NULL
#opal.assign.data(o, 'LASI_ds_w1', LASI_ds_w1)
#opal.symbol_import(o,'LASI_ds_w1', project='_Athlos_Harm_Dataset')
education <- LASI_ds_w1
save(education, file = "../RData/w1/education.RData")
rm(education)

LASI_m_ds_w1 <- LASI_ds_w1
LASI_m_ds_w1$education <- car::recode(LASI_m_ds_w1$education, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'LASI_m_ds_w1', LASI_m_ds_w1)
#opal.symbol_import(o,'LASI_m_ds_w1', project='_Athlos_Harm_Dataset')
```


# Quality estimation

No comments

```{r closeRsession, echo=FALSE,  echo=FALSE} 
opal.logout(o)
```
