==> LASI/1_Sociodemo_econom/9_nonpaid_work.Rmd <==
---
title: "Non paid work"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 
<!-- # Installation, loading packages -->
```{r setup, include=FALSE}
source("../../setup_iago.r")
```
<!-- ########################################################## --> 
<!-- # Loading DS description -->
```{r main, child='../../_DS_Rmd/nonpaid_workDS.Rmd'}
```
```{r global, echo=F}
source('../../_DS_Rmd/nonpaid_workDS.R')
```



# Data process




## India


### Wave 1


#### Study-specific variable description



| **Name** |  `we002` |`we004`|`we004_other`|`we329`|
|-|-|-|-|-|
| **Label** | `work for at least on hour last week`|`best description of current work status`|`best description of current work status other`|`work for someone else or self employed`|
| **Table name**  | `LASI-Pilot_all`| `LASI-Pilot_all`|`LASI-Pilot_all`|`LASI-Pilot_all`|
| **Categories** | `1 = yes`<br/>`2 = no` | `1 = disabled: unable to work`<br/>`2 = unemployed: interested in working but unable to find a job`<br/>`3 = homemaker`<br/>`4 = other, please specify` |`...`|`1 = employed, non-agricultural wage/salary workers`<br/>`2 = self-employed`<br/>`3 = non-paid family worker`|
| **Missings** | `NA` | `NA` | `NA` |
| **Description** |  |  |  |

```{r assign w1, echo=F}

opal.assign.table.tibble(o, 'LASI_w1','LASI.LASI-Pilot_all',variables=list('we002','we004','we004_other','we329'), missings = T)
LASI_w1 <- opal.execute(o,'LASI_w1')

```

```{r local w1, echo=F}

vbl <- LASI_w1$we002
kable(Categorical_summary(vbl, missing_values = NA)[2], caption = names(LASI_w1)[2])
kable(Categorical_summary(vbl, missing_values = NA)[1], caption = "Category")
ggplot(LASI_w1, aes(x=factor(we002))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

vbl <- LASI_w1$we004
kable(Categorical_summary(vbl, missing_values = NA)[2], caption = names(LASI_w1)[3])
kable(Categorical_summary(vbl, missing_values = NA)[1], caption = "Category")
ggplot(LASI_w1, aes(x=factor(we004))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

kable(frq(LASI_w1$we004_other[which(!is.na(LASI_w1$we004_other))])[[1]][frq(LASI_w1$we004_other[which(!is.na(LASI_w1$we004_other))])[[1]][,2]>2,], align = 'c',caption = 'we004_other')

vbl <- LASI_w1$we329
kable(Categorical_summary(vbl, missing_values = NA)[2], caption = names(LASI_w1)[5])
kable(Categorical_summary(vbl, missing_values = NA)[1], caption = "Category")
ggplot(LASI_w1, aes(x=factor(we329))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

```


#### Harmonisation algorithm

To compute `nonpaid_work` from the study-specific variable it has to be recoded as follows:

* `we004 == 1 or we002 == 1 or we329 == 1 into 996`
* `we004 == 3 into 1`
* `we329 == 3 into 2`
* `otherwise into 999`

**R script:**

```{r harmo w1}

# Categorical variable harmonization
LASI_ds_w1 <- tibble(id=LASI_w1$id)
LASI_ds_w1$nonpaid_work <- rep(999,length(LASI_w1$id))
LASI_ds_w1$nonpaid_work[which(rowSums(LASI_w1[,c("we004","we002","we329")]==1,na.rm = T)>0)] <- 996
LASI_ds_w1$nonpaid_work[which(LASI_w1$we004==3)] <- 1
LASI_ds_w1$nonpaid_work[which(tolower(gsub(" ","",LASI_w1$we004_other))=="housewife")] <- 1
LASI_ds_w1$nonpaid_work[which(LASI_w1$we329==3)] <- 2

```


#### Statistics of the new harmonized variable

```{r descript w1, echo=F}

vbl <- LASI_ds_w1$nonpaid_work
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[1], caption = "Category")
kable(Categorical_summary(vbl, missing_values = miss_values_vector)[2], caption = "Type of missing")
ggplot(LASI_ds_w1, aes(x=factor(nonpaid_work))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")

```


#### Validation

```{r crosstabulation w1, echo=F}


```






## Descriptive statistics of the harmonised variable accross populations and waves

Percentages of categories in each harmonised variable:


**R script:**

```{r summ}
LASI_ds_w1$nonpaid_work <- labelled(LASI_ds_w1$nonpaid_work, labels = cat_label)
label(LASI_ds_w1$nonpaid_work) <- ds_label

LASI_m_ds_w1 <- LASI_ds_w1
LASI_m_ds_w1$nonpaid_work <- car::recode(LASI_m_ds_w1$nonpaid_work, "miss_values_vector=NA")

tt <- frq(LASI_ds_w1$nonpaid_work)[[1]][,c(1:2,4)]
tt[14,] <- c("n", "sample size", 
             length(LASI_ds_w1$nonpaid_work)
             )
names(tt)<-c("val", "label", "w1")
kable(tt)

```



# Quality estimation




<!-- ########################################################## --> 
<!-- # Save -->
```{r save, include=FALSE, echo=F}

l.LASI <- list(w1 = LASI_ds_w1)

for(index in seq_along(l.LASI)){
  nonpaid_work <- l.LASI[[index]]
  save(nonpaid_work, file = paste0(datafolder,names(l.LASI)[index],"/nonpaid_work.RData"))
}

```
<!-- ########################################################## --> 






<!-- ########################################################## -->
<!--- # Close OPAL R Session -->
```{r closeRsession, echo=FALSE}
opal.logout(o)
```
