==> LASI/1_Sociodemo_econom/13_Wealth.Rmd <==
---
title: "Wealth"
output: 
  html_document:
    theme: united
    highlight: tango
    toc: true
    toc_depth: 6
    toc_float:
      collapsed: true
      smooth_scroll: true
    number_sections: true
---

<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->



```{r setup, include=FALSE,  echo=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_laia.r")
```

<!-- ########################################################## --> 

# Description of DataSchema variable 

* Short name: `wealth`
* Variable label: `Total score based in all sources of income transformed into a comparable and normalized categoric variable.`
* Domain: `Socio-demographic and economic characteristics`
* Value type: `categorical ordinal`
* Category coding:

**Code** | **Category Label**
-------- | ------------------
1        | 1st quintile
2        | 2nd quintile
3        | 3rd quintile
4        | 4th quintile
5        | 5th quintile

* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

* Comments: `Need to pay attention to specificities like: before/after taxes or contributions. People: HH income is prioritized. When not available individual/caregiver/informant/family income is caolected. Source: Pension/family/work/government/disablity/total. all sources are taken into account. For some studies, the maximum value can be compared (and crossed by sex and age to control biases) . Time-lapse: weekly, monthly, in the last 12 months. 'Many values are missing since not all respondents have all sources of income or, within the household there is not always more members, in these cases do not confuse missing data with missing values and mean a 0 should be added to the total sum in each case.`

```{r lab, echo=FALSE}
ds_label <- "Wealth"
ds_label_all <- "Total score based in all sources of income transformed into a comparable and normalized categoric variable"
```

# Data process

## India

### Wave 1 

#### Study-specific variable description

| **Name** | `we329` | `we365` | `we424` | `we440` | `we367_1_` | `we367_2_` | `we367_3_` | `we367_4_` | `we367_5_` | `we367_6_` | 
|-|-|-|-|-|-|-|-|-|-|-|
| **Label** | `work for someone else or self employed` | `how much earned from working on job in past 12 mo after tax income` | `how much or profits constitute own income` | `average pre tax monthly income or wage from side job` | `estimated value of subsidy per month` | `estimated value of subsidy per month` | `estimated value of subsidy per month` | `estimated value of subsidy per month` | `estimated value of subsidy per month` | `estimated value of subsidy per month` | 
| **Table name** | `LASI-Pilot_all` | `LASI-Pilot_all` | `LASI-Pilot_all` | `LASI-Pilot_all` | `LASI-Pilot_all` | `LASI-Pilot_all` | `LASI-Pilot_all` | `LASI-Pilot_all` |`LASI-Pilot_all` |`LASI-Pilot_all` |
| **Categories** | `1 = employed, non-agricultural wage/salary workers` | | | | | | | | | | 
| | `2 = self-employed` | | | | | | | | | | 
| | `3 = non-paid family worker` | | | | | | | | | | 

| **Name** | `we367_10_` | `in013_5_` | `in013_2_` | `in013_4_` | `in013_16_` | `in013_3_` | `in013_7_` | `in013_1_` | `in013_6_` | `in013_avg_2_` | 
|-|-|-|-|-|-|-|-|-|-|-|
| **Label** | `estimated value of subsidy per month` | `how much income from pension` | `how much income from pension` | `how much income from pension` | `how much income from pension` | `how much income from pension` | `how much income from pension` | `how much income from pension` | `how much income from pension` | `average monthly pension` | 
| **Table name** | `LASI-Pilot_all` | `LASI-Pilot_all` | `LASI-Pilot_all` | `LASI-Pilot_all` | `LASI-Pilot_all` | `LASI-Pilot_all` | `LASI-Pilot_all` | `LASI-Pilot_all` |`LASI-Pilot_all` |`LASI-Pilot_all` |

| **Name** | `in013_avg_6_` | `in013_avg_6_` | `in013_avg_4_` | `in013_avg_7_` | `in013_avg_1_` | `in013_avg_16_` | `in013_avg_3_` | `in013_avg_5_` | `ad018` | `ad121` | 
|-|-|-|-|-|-|-|-|-|-|-|
| **Label** | `average monthly pension` | `average monthly pension` | `average monthly pension` | `average monthly pension` | `average monthly pension` | `average monthly pension` | `average monthly pension` | `average monthly pension` | `total rental income last 12 months` | `returns on these financial investments past yr` | 
| **Table name** | `LASI-Pilot_all` | `LASI-Pilot_all` | `LASI-Pilot_all` | `LASI-Pilot_all` | `LASI-Pilot_all` | `LASI-Pilot_all` | `LASI-Pilot_all` | `LASI-Pilot_all` |`LASI-Pilot_all` |`LASI-Pilot_all` |

| **Name** | `ag023` | `ag031_income` | `ag033` | `in108a` | `in108b` | `in108c` | `in108d` | `in108e` | `in108f` | `in109a` | `in127` | `fs403` |
|-|-|-|-|-|-|-|-|-|-|-|-|-|
| **Label** | `value of crops forestry and fishing past yr` | `rental income from renting out livestock` | `value of livestock products sold` | `rice` | `wheat` | `wheat` | `kerosene` | `cooking oil` | `others` | `amt from reforestation` | `value of gifts last 12 months` | `total value of financial help received from family in last 12 months` |
| **Table name** | `LASI-Pilot_all` | `LASI-Pilot_all` | `LASI-Pilot_all` | `LASI-Pilot_all` | `LASI-Pilot_all` | `LASI-Pilot_all` | `LASI-Pilot_all` | `LASI-Pilot_all` |`LASI-Pilot_all` |`LASI-Pilot_all` | `LASI-Pilot_all` | `LASI-Pilot_all` |
| **Description** |  | | | | | | | | | | | |


```{r assign1, echo=F}
opal.assign.table.tibble(o, 'LASI_wealth','LASI.LASI-Pilot_all',
                         variables=list('we329', 'we365', 'we424', 'we440', 'we367_1_', 'we367_2_', 'we367_3_', 'we367_4_', 'we367_5_', 'we367_6_', 'we367_10_', 'in013_5_', 'in013_2_', 'in013_4_', 'in013_16_', 'in013_3_', 'in013_7_', 'in013_1_', 'in013_6_', 'in013_avg_2_', 'in013_avg_6_', 'in013_avg_4_', 'in013_avg_7_', 'in013_avg_1_', 'in013_avg_16_', 'in013_avg_3_', 'in013_avg_5_', 'ad018', 'ad121', 'ag023', 'ag031_income', 'ag033', 'in108a', 'in108b', 'in108c', 'in108d', 'in108e', 'in108f', 'in109a', 'in127', 'fs403'), missings = TRUE)
LASI_wealth <- opal.execute(o,'LASI_wealth')

```


##### Data management

Variables with different quinds of incomes and assets are computed from the study-specific variables: 

* **R0IEARN:** Respondent's primary job wage/salary income, self-employment income, and 2nd job earnings from the last 12 months and is not included in the computation of household income
    + `R0IEARN = we365 + we424 + we440 + we367`
* **R0IPEN:** sum of respondent's income from all pension
    + `R0IPEN = in013`
* **H0ICAP:** the sum of household net income from agricultural activities and income from all assets, such as business or agricultural income, gross rent, dividend and interest income, or other asset income.
    + `H0ICAP = ad018 + ad121 + ag023 + ag031_income + ag033`
* **H0IGXFR:** the income from government transfers including ration
    + `H0IGXFR = in108a + in108b + in108c + in108d + in108e + in108f + in109a`
* **in127:**  the income from remittances
* **fs403:** the income from private transfers e.g., gifts, donations, and inheritances


```{r d_m}
LASI_wealth$we367_1_ <- LASI_wealth$we367_1_*12
LASI_wealth$we367_2_ <- LASI_wealth$we367_2_*12
LASI_wealth$we367_3_ <- LASI_wealth$we367_3_*12
LASI_wealth$we367_4_ <- LASI_wealth$we367_4_*12
LASI_wealth$we367_5_ <- LASI_wealth$we367_5_*12
LASI_wealth$we367_6_ <- LASI_wealth$we367_6_*12
LASI_wealth$we367_10_ <- LASI_wealth$we367_10_*12


LASI_wealth$R0IEARN <- apply(X = select(LASI_wealth, we365, we424, we440, we367_1_, we367_2_, we367_3_, we367_4_, we367_5_, we367_6_, we367_10_), MARGIN = 1, FUN = function(x) sum(x, na.rm = TRUE))
LASI_wealth$R0IEARN <- car::recode(LASI_wealth$R0IEARN, "0=NA")


LASI_wealth$in013_5_[is.na(LASI_wealth$in013_5_)] <- LASI_wealth$in013_avg_5_[is.na(LASI_wealth$in013_5_)]*12
LASI_wealth$in013_16_[is.na(LASI_wealth$in013_16_)] <- LASI_wealth$in013_avg_16_[is.na(LASI_wealth$in013_16_)]*12
LASI_wealth$in013_3_[is.na(LASI_wealth$in013_3_)] <- LASI_wealth$in013_avg_3_[is.na(LASI_wealth$in013_3_)]*12
LASI_wealth$in013_7_[is.na(LASI_wealth$in013_7_)] <- LASI_wealth$in013_avg_7_[is.na(LASI_wealth$in013_7_)]*12
LASI_wealth$in013_1_[is.na(LASI_wealth$in013_1_)] <- LASI_wealth$in013_avg_1_[is.na(LASI_wealth$in013_1_)]*12
LASI_wealth$in013_6_[is.na(LASI_wealth$in013_6_)] <- LASI_wealth$in013_avg_6_[is.na(LASI_wealth$in013_6_)]*12
LASI_wealth$in013_2_[is.na(LASI_wealth$in013_2_)] <- LASI_wealth$in013_avg_2_[is.na(LASI_wealth$in013_2_)]*12
LASI_wealth$in013_4_[is.na(LASI_wealth$in013_4_)] <- LASI_wealth$in013_avg_4_[is.na(LASI_wealth$in013_4_)]*12


LASI_wealth$R0IPEN <- apply(X = select(LASI_wealth, in013_5_, in013_2_, in013_4_, in013_16_, in013_3_, in013_7_, in013_1_, in013_6_), MARGIN = 1, FUN = function(x) sum(x, na.rm = TRUE))
LASI_wealth$R0IPEN <- car::recode(LASI_wealth$R0IPEN, "0=NA")

LASI_wealth$H0ICAP <- apply(X = select(LASI_wealth, ad018, ad121, ag023, ag031_income, ag033), MARGIN = 1, FUN = function(x) sum(x, na.rm = TRUE))
LASI_wealth$H0ICAP <- car::recode(LASI_wealth$H0ICAP, "0=NA")

LASI_wealth$H0IGXFR <- apply(X = select(LASI_wealth, in108a, in108b, in108c, in108d, in108e, in108f, in109a), MARGIN = 1, FUN = function(x) sum(x, na.rm = TRUE))
LASI_wealth$H0IGXFR <- car::recode(LASI_wealth$H0IGXFR, "0=NA")
```


```{r local1, echo=F}
vari <- LASI_wealth$R0IEARN
kable(Continuous_summary(vari, missing_values = NA)[3], caption = "Year of birth") 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(LASI_wealth, aes(vari)) + geom_histogram(stat="count", fill="steelblue") + xlab("Year of birth") + ylab("Frequency")

vari <- LASI_wealth$R0IPEN
kable(Continuous_summary(vari, missing_values = NA)[3], caption = "Year of birth") 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(LASI_wealth, aes(vari)) + geom_histogram(stat="count", fill="steelblue") + xlab("Year of birth") + ylab("Frequency")

vari <- LASI_wealth$H0ICAP
kable(Continuous_summary(vari, missing_values = NA)[3], caption = "Year of birth") 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(LASI_wealth, aes(vari)) + geom_histogram(stat="count", fill="steelblue") + xlab("Year of birth") + ylab("Frequency")

vari <- LASI_wealth$H0IGXFR
kable(Continuous_summary(vari, missing_values = NA)[3], caption = "Year of birth") 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(LASI_wealth, aes(vari)) + geom_histogram(stat="count", fill="steelblue") + xlab("Year of birth") + ylab("Frequency")

vari <- LASI_wealth$in127
kable(Continuous_summary(vari, missing_values = NA)[3], caption = "Year of birth") 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(LASI_wealth, aes(vari)) + geom_histogram(stat="count", fill="steelblue") + xlab("Year of birth") + ylab("Frequency")

vari <- LASI_wealth$fs403
kable(Continuous_summary(vari, missing_values = NA)[3], caption = "Year of birth") 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(LASI_wealth, aes(vari)) + geom_histogram(stat="count", fill="steelblue") + xlab("Year of birth") + ylab("Frequency")
```


#### Harmonisation algorithm
To compute the harmonized variable from the study specific varibale sum up all the variables and divide the population into 5 groups each containing 20$\%$ of the corresponding N and the missing values have to be recoded as follows: 

* `NA into 999`

**R script:**

```{r harmo1}
LASI_wealth$sum <- apply(X = select(LASI_wealth, R0IEARN, R0IPEN, H0ICAP, H0IGXFR, in127, fs403), MARGIN = 1, FUN = function(x) sum(x, na.rm = TRUE))
LASI_wealth$sum <- car::recode(LASI_wealth$sum, "0=NA")

qq <- quantile(LASI_wealth$sum, probs = seq(0, 1, .2), na.rm = TRUE)

LASI_ds_w1 <- tibble(id=LASI_wealth$id)
LASI_ds_w1$wealth <- car::recode(LASI_wealth$sum, "0:qq[2]=1; (qq[2]+1):qq[3]=2; (qq[3]+1):qq[4]=3; (qq[4]+1):qq[5]=4; (qq[5]+1):hi=5")
LASI_ds_w1$wealth <- car::recode(LASI_ds_w1$wealth, "NA=999")
LASI_ds_w1$wealth <- as.integer(LASI_ds_w1$wealth)
LASI_ds_w1$wealth <- labelled(LASI_ds_w1$wealth, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript1, echo=F}
vari <- LASI_ds_w1$wealth
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label) 
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[1], caption = "Category")
kable(Categorical_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Type of missing")
ggplot(LASI_ds_w1, aes(x=factor(wealth))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Validation

As some individuals are non-paid family worker their low or null income doesn't mean they don't have subsitance economy. They are listed below

```{r crosstabulation1, echo=F}
LASI_wealth <- full_join(LASI_wealth, LASI_ds_w1, by='id')
kable(LASI_wealth[which(LASI_wealth$we329==3), c(1,32, 47, 48)])
```



```{r importOPAL1, echo=F}
#opal.assign.data(o, 'LASI_ds_w1', LASI_ds_w1)
#opal.symbol_import(o,'LASI_ds_w1', project='_Athlos_Harm_Dataset')
wealth <- LASI_ds_w1
save(wealth, file = "../RData/w1/wealth.RData")
rm(wealth)

LASI_m_ds_w1 <- LASI_ds_w1
LASI_m_ds_w1$wealth <- car::recode(LASI_m_ds_w1$wealth, "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
#opal.assign.data(o, 'LASI_m_ds_w1', LASI_m_ds_w1)
#opal.symbol_import(o,'LASI_m_ds_w1', project='_Athlos_Harm_Dataset')
```


# Quality estimation
Household income form job, fringe benefit, side job, rental income, finantial investments, personal loans, 
renting livestock, income from products, remittances, private transfers, gifts.
The continuous variables were sumed up and quintiles were created based on the household income of each 
participant that divided the population into 5 groups each containing 20% of the corresponding N.   


```{r closeRsession, echo=FALSE}
opal.logout(o)
```
